pyproject.toml
[project]
name = "frcm"
version = "0.1.0"
description = ""
maintainers = [
    { "name" = "Lars Michael Kristensen", "email" = "lmkr@users.noreply.github.com" },
    { "name" = "Patrick Stünkel", "email" = "webminz@users.noreply.github.com" }
]
readme = "README.md"
requires-python = ">= 3.11"
dependencies = [
    "pydantic >=2.10",
    "python-dateutil == 2.8.2",
    "pytest == 7.4.4",
    "pytest-mock == 3.12.0",
    "numpy >=2.1",
    "requests == 2.31.0",
    "python-decouple == 3.8",
    "pydantic-core>=2.14.6",
    "fastapi >=0.115.0",
    "uvicorn >=0.30.0",
    "fastapi >= 0.115.0",
    "uvicorn[standard] >= 0.32.0",
    "python-multipart >= 0.0.22",
    "httpx >= 0.24.0",
    "paho-mqtt >= 1.6.1",
]

[project.scripts]
frcm = "frcm:console_main"
frcm-api = "frcm.api.server:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
